jsoned/keywords/required_properties_keyword.py
from functools import partial

from jsoned.errors.schema_parse_error import SchemaParseError
from jsoned.json_core import AssertionKeyword, JsonSchema
from jsoned.types import JsonObject, JsonType
from jsoned.validators.core_validators import ValidatorsMap
from jsoned.validators.object_validators import validate_required_properties


class RequiredPropertiesKeyword(AssertionKeyword):
    key = "required"

    def apply(self, document: JsonSchema, node: JsonObject, validator: ValidatorsMap):
        if node[self.key].type != JsonType.ARRAY:
            raise SchemaParseError.for_invalid_keyword_value(node, self.key, JsonType.ARRAY)

        validator[self.key] = partial(
            validate_required_properties,
            expected_properties=[str(item) for item in node[self.key]]
        )
